LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-y-in-scrolling-div-expected.txt

FAIL div.scrollTop should be 322. Was 0.
PASS div.scrollLeft is window.expectedScrollLeft
